@@ -1,4 +1,8 @@
 ﻿
+using DotNETDepends.Output;
+using Microsoft.Build.Locator;
+using System.Text.Json;
+
 namespace DotNETDepends;
 class Program
 {
@@ -11,10 +15,25 @@ class Program
     {
         if (args.Length == 1 && File.Exists(args[0]))
         {
-            var reader = new SolutionReader();
-            await reader.ReadSolutionAsync(args[0]).ConfigureAwait(false);
-            //Set this to indicate success
-            Environment.ExitCode = 0;
+            AnalysisOutput output = new();
+            try
+            {
+                //This finds the MSBuild libs from the .NET SDK
+                //It can only be called once without throwing
+                MSBuildLocator.RegisterDefaults();
+
+                var reader = new SolutionReader();
+                await reader.ReadSolutionAsync(args[0], output).ConfigureAwait(false);
+                //Set this to indicate success
+                Environment.ExitCode = 0;
+            }catch(Exception ex)
+            {
+                output.AddErrorMessage(ex.ToString());
+            }
+            Console.WriteLine("__codesee.output.begin__");
+            var options = new JsonSerializerOptions { WriteIndented = true };
+            string jsonString = JsonSerializer.Serialize(output, options);
+            Console.WriteLine(jsonString);
         }
         else
         {
