@@ -14,19 +14,22 @@ namespace Disassembler
     {
         private readonly SourceTypeBuilder builder;
         //Map of types in a namespace
-        private readonly Dictionary<string, List<SourceType>> types = new();
+        private readonly Dictionary<string, HashSet<SourceType>> types = new();
 
         /**
          * Gets all of the types we have added during disassembly
          */
-        public List<SourceType> SourceTypes
+        public HashSet<SourceType> SourceTypes
         {
             get
             {
-                List<SourceType> result = new();
-                foreach (List<SourceType> typeList in types.Values)
+                HashSet<SourceType> result = new();
+                foreach (HashSet<SourceType> typeList in types.Values)
                 {
-                    result.AddRange(typeList);
+                    foreach(var type in typeList)
+                    {
+                        result.Add(type);
+                    }
                 }
                 return result;
             }
@@ -38,6 +41,17 @@ namespace Disassembler
 
         }
 
+        private void AddTypeToNamespace(string ns, SourceType type)
+        {
+            types.TryGetValue(ns, out HashSet<SourceType>? nsTypes);
+            if (nsTypes == null)
+            {
+                nsTypes = new HashSet<SourceType>();
+                types.Add(ns, nsTypes);
+            }
+            nsTypes.Add(type);
+        }
+
         /**
          * Constructs the source type using the abstract SourceTypeBuilder
          * It then adds it to the types map by namespaces
@@ -45,13 +59,11 @@ namespace Disassembler
         public void AddSource(string filePath, string projectDir, string solutionDir, string rootNamepace)
         {
             var type = builder.Build(rootNamepace, filePath, projectDir, solutionDir);
-            types.TryGetValue(type.Namespace, out List<SourceType>? nsTypes);
-            if (nsTypes == null)
+            AddTypeToNamespace(type.Namespace, type);
+            if(type.NamespaceAlias!= null)
             {
-                nsTypes = new List<SourceType>();
-                types.Add(type.Namespace, nsTypes);
+                AddTypeToNamespace(type.NamespaceAlias, type);  
             }
-            nsTypes.Add(type);
         }
 
         /**
@@ -60,7 +72,7 @@ namespace Disassembler
          */
         public SourceType? Locate(Type type)
         {
-            types.TryGetValue(type.Namespace ?? "", out List<SourceType>? nsTypes);
+            types.TryGetValue(type.Namespace ?? "", out HashSet<SourceType>? nsTypes);
             if (nsTypes != null)
             {
                 foreach (SourceType nsType in nsTypes)
